requirements.txt
streamlit~=1.28.1
Pillow~=10.1.0
pydeck
pandas~=2.1.1
torch~=2.1.0
opencv-python~=4.8.1.78
torchvision~=0.16.0